refactor(meeting): DTO 클래스 이름 변경 및 패키지 구조 업데이트
