+++ src/main/java/com/eventitta/meeting/dto/response/MeetingDetailResponse.java
@@ -1,19 +1,20 @@
-package com.eventitta.meeting.dto;
+package com.eventitta.meeting.dto.response;
 
 import com.eventitta.meeting.domain.MeetingStatus;
 import io.swagger.v3.oas.annotations.media.Schema;
 
 import java.time.LocalDateTime;
+import java.util.List;
 
-@Schema(description = "모임 목록 조회 응답 아이템")
-public record MeetingSummaryResponse(
+@Schema(description = "모임 상세 정보 응답")
+public record MeetingDetailResponse(
     @Schema(description = "모임 ID", example = "1")
     Long id,
 
     @Schema(description = "모임 제목", example = "함께하는 스프링 스터디")
     String title,
 
-    @Schema(description = "모임 설명 요약", example = "매주 주말에 만나 스프링 심화 내용을 공부합니다.")
+    @Schema(description = "모임 설명", example = "매주 주말에 만나 스프링 심화 내용을 공부합니다.")
     String description,
 
     @Schema(description = "모임 시작 시간", example = "2025-08-01T10:00:00")
@@ -46,7 +47,10 @@ public record MeetingSummaryResponse(
     @Schema(description = "모임장 닉네임", example = "스프링마스터")
     String leaderNickname,
 
-    @Schema(description = "거리 (km, 위치 기반 검색 시에만 값이 있음)", example = "3.2")
-    Double distance
+    @Schema(description = "모임장 프로필 이미지 URL", example = "https://example.com/profile.jpg")
+    String leaderProfileUrl,
+
+    @Schema(description = "참여자 목록")
+    List<ParticipantResponse> participants
 ) {
 }
